WEBNEW-214 :lipstick: Moved bespoke styling into Blog styles
