+++ src/blog/article/BlogArticlePage.styles.ts
@@ -11,8 +11,13 @@ import {
 import { RichText } from '../../components/richText'
 import { fontSizes } from '../../theme/fonts'
 import { mediaQueries } from '../../theme/mediaQueries'
+import { spacing } from '../../theme/space'
 
 export const MAX_CONTENT_WIDTH = 624
+export const CONTENT_PADDING_X = {
+  default: 'lg',
+  md: 'xxl',
+} as const
 
 export const InfoWrapper = styled.div<FlexboxProps & LayoutProps & SpaceProps>`
   ${compose(flexbox, layout, space)}
@@ -22,9 +27,20 @@ export const TitleWrapper = styled.div<FlexboxProps & LayoutProps & SpaceProps>`
   ${compose(flexbox, layout, space)}
 `
 
-export const Content = styled(RichText)`
+export const Content = styled(RichText).attrs({ paddingX: CONTENT_PADDING_X })`
   > p:first-of-type {
     font-size: ${fontSizes.body_lg};
+  }
+
+  /* RippedSection full-width image :( */
+  > section {
+    margin-left: -${spacing[CONTENT_PADDING_X.default]};
+    margin-right: -${spacing[CONTENT_PADDING_X.default]};
+
+    ${mediaQueries.md} {
+      margin-left: -${spacing[CONTENT_PADDING_X.md]};
+      margin-right: -${spacing[CONTENT_PADDING_X.md]};
+    }
   }
 
   ${mediaQueries.md} {
